Introduce PCA Visualisations (EDA) and GridSearch to improve model.
This commit introduces a few modifications. First, dimensionality reduction using PCA to produce some visualisations that demonstrate the relatively separated nature of the data meaning simple models would likely suffice. Second, the use of GridSearch on the RandomForest model to improve accuracy by over 3%. Finally, a fix in the streamlit app that resolves the issue of the model not producing a position prediction by increasing height by 0.01.