Built model for determining most similar active NBA player.
This commit introduces a a model that takes the user input and finds the most similar active NBA player, with the model being integrated in the streamlit app. Furthermore, there are player images uploaded in a folder that will be used as part of the output of the most similar active NBA player.