Implement new player similarity model.
This commit introduces a new player similarity model based on a players statistics. Instead of using cosine similarity (which considers vector direction and not magnitude), we use a more appropriate euclidean distance (which also considers vector magnitude). A prior issue was the suggestion of players with similar proportions of stats (such as 2:1 points:assists), but signficantly differing magnitudes. This new model resolves this and is more contextually appropriate.